+++ tests/JWT.Tests.Common/JwtDecoderTest.cs
@@ -1,7 +1,8 @@
 using System;
 using JWT.Algorithms;
 using JWT.Serializers;
-using JWT.Tests.Common;
+using JWT.Tests.Common.Internal;
+using JWT.Tests.Common.Models;
 using Xunit;
 
 namespace JWT.Tests
